Moved Product.cs to DataLayer
